[WIP] Proposal to establish `burn::baselib::...` collecting pre-tensor common libs; starting with indexing.
