+++ crates/burn-common/src/baselib/mod.rs
@@ -1,0 +1,2 @@
+/// The indexing module.
+pub mod indexing;
