+++ crates/burn-tensor/src/tensor/api/mod.rs
@@ -9,17 +9,16 @@ mod float;
 mod int;
 mod kind;
 mod numeric;
-mod slice;
 mod sort;
 mod transaction;
 
 pub use argwhere::argwhere_data;
 pub use autodiff::*;
 pub use base::*;
+pub use burn_common::baselib::indexing::*;
 pub use cartesian_grid::cartesian_grid;
 pub use float::{DEFAULT_ATOL, DEFAULT_RTOL};
 pub use kind::*;
 pub use numeric::*;
-pub use slice::*;
 pub use sort::{argsort, sort, sort_with_indices};
 pub use transaction::*;
