+++ crates/burn-tensor/src/tests/quantization/ops/slice.rs
@@ -1,8 +1,9 @@
 #[burn_tensor_testgen::testgen(q_slice)]
 mod tests {
     use super::*;
+    use burn_common::s;
     use burn_tensor::TensorData;
-    use burn_tensor::{Tolerance, ops::FloatElem, s};
+    use burn_tensor::{Tolerance, ops::FloatElem};
     type FT = FloatElem<TestBackend>;
 
     #[test]
